added routes, blade views, fixed migrations/seeders/factories

+++ lottery.sh
@@ -1,21 +1,16 @@
 
-sudo chmod -R 777 ./storage
-if [ $1 = "up" ]; then
-    sudo docker-compose up -d --build
+if [ -z "$#" ]
+then
+      echo "NO ARGUMENTS"
+else
+    if [ $1 = "up" ]; then
+        sudo docker-compose up -d --build
+    elif [ $1 = "down" ]; then
+        sudo docker-compose down
+    elif [ $1 = "status" ]; then
+        sudo docker container ls
+    else
+        sudo docker-compose run --rm $1 $2 $3 $4
+    fi
 fi
-
-if [ $1 = "down" ]; then
-    sudo docker-compose down
-fi
-
-if [ $1 = "status" ]; then
-    sudo docker container ls
-fi
-
-if [ $1 = "artisan" ]; then
-    sudo docker-compose run --rm artisan $2 $3 $4
-fi
-
-if [ $1 = "composer" ]; then
-    sudo docker-compose run --rm composer $2 $3 $4
-fi
+sudo chmod -R 777 ./
